refactor: refactor content component
Dashboard has no more responsible to control contexts used by Content component
